+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [Autocompletion for Column names in table expressions][13848]
 - [When connecting to port a value with additional type, a necessary type cast
   is included][14028]
+- [Function docs in autocomplete in table expressions][14059]
 
 [13685]: https://github.com/enso-org/enso/pull/13685
 [13658]: https://github.com/enso-org/enso/pull/13658
@@ -22,6 +23,7 @@
 [13848]: https://github.com/enso-org/enso/pull/13848
 [14028]: https://github.com/enso-org/enso/pull/14028
 [13976]: https://github.com/enso-org/enso/pull/13976
+[14059]: https://github.com/enso-org/enso/pull/14059
 
 #### Enso Standard Library